+++ Camel.Engine/Tokens/DivideOperatorToken.cs
@@ -2,7 +2,7 @@ namespace Lepecki.Playground.Camel.Engine.Tokens
 {
     public class DivideOperatorToken : BinaryOperatorToken
     {
-        protected override double Calculate(double a, double b)
+        protected override decimal Calculate(decimal a, decimal b)
         {
             return a / b;
         }
